requirements.txt
flask-wtf
flask-login
flask-sqlalchemy
flask-migrate
flask-login
email-validator
pydenticon